refactor: move update handler to Backend module
