+++ Moc/MocApp+Injection.swift
@@ -59,108 +59,7 @@ extension Resolver {
 
     // swiftlint:disable cyclomatic_complexity function_body_length
     public static func registerBackend() {
-        let tdApi = TdApi(client: TdClientImpl(completionQueue: .global(), logger: TdLogger()))
-
-        Task {
-            #if DEBUG
-            try? await tdApi.setLogVerbosityLevel(newVerbosityLevel: 5)
-            #else
-            try? await tdApi.setLogVerbosityLevel(newVerbosityLevel: 0)
-            #endif
-        }
-        tdApi.client.run {
-            do {
-                let update = try tdApi.decoder.decode(Update.self, from: $0)
-                switch update {
-                        // MARK: - Authorization state
-                    case .updateAuthorizationState(let state):
-                        switch state.authorizationState {
-                            case .authorizationStateWaitTdlibParameters:
-                                SystemUtils.post(notification: .authorizationStateWaitTdlibParameters)
-                                Task {
-                                    try? await tdApi.setTdlibParameters(parameters: TdlibParameters(
-                                        apiHash: Secret.apiHash,
-                                        apiId: Secret.apiId,
-                                        applicationVersion: (
-                                            Bundle.main.infoDictionary?["CFBundleShortVersionString"] as? String
-                                        ) ?? "Unknown",
-                                        databaseDirectory: "td",
-                                        deviceModel: SystemUtils.macModel,
-                                        enableStorageOptimizer: true,
-                                        filesDirectory: "td",
-                                        ignoreFileNames: false,
-                                        systemLanguageCode: "en-US",
-                                        systemVersion: SystemUtils.osVersionString,
-                                        useChatInfoDatabase: true,
-                                        useFileDatabase: true,
-                                        useMessageDatabase: true,
-                                        useSecretChats: false,
-                                        useTestDc: false
-                                    ))
-                                }
-                            case .authorizationStateWaitEncryptionKey(let info):
-                                SystemUtils.post(notification: .authorizationStateWaitEncryptionKey, withObject: info)
-                                Task {
-                                    try? await tdApi.checkDatabaseEncryptionKey(
-                                        encryptionKey: MocApp.tdDatabaseEncryptionKey
-                                    )
-                                }
-                            case .authorizationStateWaitPhoneNumber:
-                                SystemUtils.post(notification: .authorizationStateWaitPhoneNumber)
-                            case .authorizationStateWaitCode(let info):
-                                SystemUtils.post(notification: .authorizationStateWaitCode, withObject: info)
-                            case .authorizationStateWaitRegistration(let info):
-                                SystemUtils.post(notification: .authorizationStateWaitRegistration, withObject: info)
-                            case .authorizationStateWaitPassword(let info):
-                                SystemUtils.post(notification: .authorizationStateWaitPassword, withObject: info)
-                            case .authorizationStateReady:
-                                Task {
-                                    do {
-                                        _ = try await tdApi.loadChats(chatList: .chatListMain, limit: 15)
-                                        _ = try await tdApi.loadChats(chatList: .chatListArchive, limit: 15)
-                                    } catch {
-                                        logger.error("Failed to load chats")
-                                    }
-                                }
-                                SystemUtils.post(notification: .authorizationStateReady)
-                            case .authorizationStateWaitOtherDeviceConfirmation(let info):
-                                SystemUtils.post(
-                                    notification: .authorizationStateWaitOtherDeviceConfirmation,
-                                    withObject: info
-                                )
-                            case .authorizationStateLoggingOut:
-                                SystemUtils.post(notification: .authorizationStateLoggingOut)
-                            case .authorizationStateClosing:
-                                SystemUtils.post(notification: .authorizationStateClosing)
-                            case .authorizationStateClosed:
-                                SystemUtils.post(notification: .authorizationStateClosed)
-                        }
-                        // MARK: - Chat updates
-                    case .updateChatPosition(let info):
-                        SystemUtils.post(notification: .updateChatPosition, withObject: info)
-                    case .updateNewMessage(let info):
-                        SystemUtils.post(notification: .updateNewMessage, withObject: info)
-                    case .updateChatLastMessage(let info):
-                        SystemUtils.post(notification: .updateChatLastMessage, withObject: info)
-                    case .updateNewChat(let info):
-                        SystemUtils.post(notification: .updateNewChat, withObject: info)
-                    case .updateFile(let info):
-                        SystemUtils.post(notification: .updateFile, withObject: info)
-
-                    default:
-                        #if DEBUG
-                        logger.warning("Unhandled TDLib update \(update)")
-                        #endif
-                }
-            } catch {
-                #if DEBUG
-                logger.error("Error in TDLib update handler \(error.localizedDescription)")
-                #endif
-            }
-        }
-        register { tdApi }
-            .scope(.shared)
-        register { TdChatDataSource() as ChatDataSourcable }
+        register { TdChatDataSource() as ChatDataSource }
             .scope(.shared)
     }
 }
@@ -169,24 +68,13 @@ extension Resolver {
 @main
 struct MocApp: App {
     @NSApplicationDelegateAdaptor var appDelegate: AppDelegate
-
-    static var tdDatabaseEncryptionKey: Data {
-        let keychain = KeychainSwift()
-        let encryptionKey = keychain.getData("tdDatabaseEncryptionKey")
-        if encryptionKey == nil {
-            let key = SymmetricKey(size: .bits256).withUnsafeBytes {
-                return Data(Array($0))
-            }
-            keychain.set(key, forKey: "tdDatabaseEncryptionKey", withAccess: .accessibleAfterFirstUnlock)
-            return key
-        } else {
-            return encryptionKey!
-        }
-    }
+    private let logger = Logging.Logger(label: "TDLibUpdates")
 
     init() {
         Resolver.registerUI()
         Resolver.registerBackend()
+        TdApi.shared.append(TdApi(client: TdClientImpl(completionQueue: .main, logger: TdLogger())))
+        TdApi.shared[0].startTdLibUpdateHandler()
     }
 
     var body: some Scene {
